+++ Schema/src/binary/attributes/memory/IAtPositionAttribute.cs
@@ -1,6 +1,6 @@
 ﻿namespace schema.binary.attributes {
   public interface IAtPositionAttribute {
     string OffsetName { get; }
-    int? NullValue { get; }
+    long? NullValue { get; }
   }
 }
