+++ Schema/src/binary/attributes/memory/IPointerToAttribute.cs
@@ -1,6 +1,6 @@
 ﻿namespace schema.binary.attributes {
   public interface IPointerToAttribute {
     IChain<IAccessChainNode> AccessChainToOtherMember { get; }
-    int? NullValue { get; }
+    long? NullValue { get; }
   }
 }
