+++ Schema/src/binary/attributes/memory/RAtPositionOrNullAttribute.cs
@@ -3,9 +3,9 @@
 
 namespace schema.binary.attributes {
   [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-  public class RAtPositionOrNullAttribute(string offsetName, int nullValue = 0)
+  public class RAtPositionOrNullAttribute(string offsetName, long nullValue = 0)
       : Attribute, IAtPositionAttribute {
     public string OffsetName { get; } = offsetName;
-    public int? NullValue { get; } = nullValue;
+    public long? NullValue { get; } = nullValue;
   }
 }
